+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [4.8.0](https://github.com/deltaDNA/unity-sdk/releases/tag/4.8.0) (2018-??-??)
+### Added
+- Test ads support for when running in the Unity Editor.
+
 ## [4.7.1](https://github.com/deltaDNA/unity-sdk/releases/tag/4.7.1) (2018-02-06)
 ### Fixed
 - Disabled chunked transfer for Unity 2017.3.